Scope aspect name uniqueness to a Structure; drop global registries
AspectKey names no longer need to be globally unique — only unique
within a single Structure's declared aspect set. Different composites
can each carry an aspect named "Id", "Status", etc., backed by distinct
AspectKey instances.

The aspects-serialization strategy moves out of Composite (where it had
to use module-level Map<string,AspectKey> / Map<symbol,AspectKey>
lookups) and into the Structure factory, where it walks the declared
refs directly. Name/symbol resolution becomes a closure-local concern.
The Composite base class is back to a plain Map<symbol, unknown> with
no @Property.

Side benefit: the Structure factory now enforces same-Structure name
uniqueness explicitly at construction time, throwing
'Structure has duplicate aspect name "X"' if two refs collide.
